slow "time to first byte" on logged in state
When the user is logged in the browser measure for time to first byte is very slow (1.5s). I am testing with a remote mongodb which might contribute to the issue. In that situation also responses for js-files, assets and css all suffer the same initial delay. My debugging brought me to the conclusion that the handleOnAllRoutes has some side effect that is causing this. I do not see a reason why handleOnAllRoutes would need to run on every request and especially for static files / assets?
Therefore i wrapped the call to handleOnAllRoutes into the picker route when we are anticipating an html rendering to occur.